Insert recommendation node by overriding popup bindings
Rather than inserting a new popup, we can use CSS specificity to
override the existing popup's XBL bindings, inserting our hbox above the
list of results.

Holding off on adding UI classes in this commit, so we still aren't
injecting any scripts into each window.

Using id instead of anonid in the recommendation element, so that we can
simply do `document.getElementById` to grab a pointer to it.

Closes #3.