Show recommendations in response to 'paste' events on the urlbar
Note: trying to listen for 'cut' events on the urlbar didn't work at
all. It seems fairly unusual for part of the urlbar contents to be cut,
so rather than look for a comprehensive fix, just deciding to stick with
the simple, much more common case of paste support.

Fixes #68.