Adjust XUL popup state, fixing popup in Nightly
First, when the popup opens for the first time, its `mInput` will be
null, which breaks rendering of all the results. To work around this,
detect when the popup opens for the first time, via a MutationObserver
on the popup's 'hidden' attribute, then set the missing `mInput` state.

Second, each time the popup closes, its `mPopupOpen` state needs to be
reset to false. This is normally done by the `popuphiding` listener
defined in XBL. However, because our code defines its own `popuphiding`
listener, it seems that the XBL handler is not called. Work around this
by temporarily detaching our listener, firing a synthetic `popuphiding`
event on the XUL popup element, then reattaching the listener.

Note that I'm not convinced attaching a DOM Level 2 event listener
should ever have the effect of breaking an XBL (pseudo-DOM Level 0)
handler. Treat the explanation in the previous paragraph as a
hypothesis, not as a theorem :-P

Fixes #138.