Bind functions before passing to WindowWatcher
Perhaps for cross-window security sandboxing, the value of `ww._thisArg`
is converted to `undefined` when passing from the WindowWatcher context
to the individual window context.

To work around this limitation, bind the WindowWatcher callbacks in the
individual window code, rather than using Function.prototype.call() to
dynamically assign the `this` value from the WindowWatcher context.

Also, improve documentation of some complex code in the WindowWatcher.

Fixes #145.