Detect mouse detection via mousemove, not mouseover
Effective highlight stealing requires adjusting the highlight over many
(currently 6) successive turns for each DOM Mutation observed as results
stream into the popup. Because the stealing spans 6 * 17ms ~= 100ms, we
need to detect when the mouse is moving and stop stealing the highlight
immediately. We used to do this by listening for mouseover events on the
popup, then canceling further highlight stealing until the popup
contents were updated.

When the popup opens under the mouse pointer, one or more spurious
mouseover events are fired. This makes highlight stealing buggy: we can
ignore a single mouseover event, but there's no easy way to know how
many mouseovers will be fired when the DOM is updated with new results.

In contrast, mousemove events are not spuriously fired when the popup is
opened under the mouse pointer. Mousemove events fire even more often
than mouseover events, so they are still a great tool to detect when the
user is actively mousing through the popup (the case where we want to
abort highlight stealing, even if results are still streaming in).

* Rename mouseover listener / variable names to mousemove.

* Remove the 'ignoreNextMouseOver' state on the popup; it's not needed,
  now that we are using mousemove.

* Update code docs.

Fixes #143.