Disable one-off search buttons and simplify popup initialization
This commit contains a small cluster of changes that jointly fix #286:

Connecting the urlbar to the popup in an XBL constructor works far
better than the old onFirstPopupOpen approach. This is a better way of
fixing #138, and it resolves the 'maxRows - 1 is undefined' error.

Disabling one-off search by disabling the pref and removing the one-off
search buttons from the XUL DOM (via hidden=true) closes #292,
closes #293, closes #298--but those bugs aren't really fixed, just
hidden. Filed #304 to re-enable one-off search as time allows. Note that
the disabletab attribute was added just to make the set of attributes in
our one-off-search-buttons element match the element in FF.)

Switching our popup.js listener from the 'popupshowing' event to the
'popupshown' event fixes #295, with some cascading changes related to
renaming the function and the signal fired by the function.

\o/